:construction: [#550] Improvements tests, client
