+++ src/objects/core/models.py
@@ -15,7 +15,7 @@ from zgw_consumers.models import Service
 from objects.utils.client import get_objecttypes_client
 
 from .query import ObjectQuerySet, ObjectRecordQuerySet, ObjectTypeQuerySet
-from .utils import check_objecttype
+from .utils import check_objecttype_cached
 
 
 class ObjectType(models.Model):
@@ -53,14 +53,13 @@ class ObjectType(models.Model):
         if exclude and "service" in exclude:
             return
 
-        client = get_objecttypes_client(self.service)
-
-        try:
-            object_type_data = client.get_objecttype(self.uuid)
-        except (requests.RequestException, ConnectionError, ValueError) as exc:
-            raise ValidationError(f"Objecttype can't be requested: {exc}")
-        except requests.exceptions.JSONDecodeError:
-            raise ValidationError("Object type version didn't have any data")
+        with get_objecttypes_client(self.service) as client:
+            try:
+                object_type_data = client.get_objecttype(self.uuid)
+            except (requests.RequestException, ConnectionError, ValueError) as exc:
+                raise ValidationError(f"Objecttype can't be requested: {exc}")
+            except requests.exceptions.JSONDecodeError:
+                raise ValidationError("Object type version didn't have any data")
 
         if not self._name:
             self._name = object_type_data["name"]
@@ -156,7 +155,7 @@ class ObjectRecord(models.Model):
         super().clean()
 
         if hasattr(self.object, "object_type") and self.version and self.data:
-            check_objecttype(self.object.object_type, self.version, self.data)
+            check_objecttype_cached(self.object.object_type, self.version, self.data)
 
     def save(self, *args, **kwargs):
         if not self.id and self.object.last_record:
